+++ epyk/core/js/packages/JsLeaflet.py
@@ -9,19 +9,55 @@ class LTooltip:
     self._js, self.__is_attached = [], False
     self.varId = "L.marker(%s)" % latlng
 
-  def openTooltip(self):
+  def openTooltip(self, latlng=None, options=None):
     """
     Description:
     -----------
+    Opens the specified tooltip.
 
+    Related Pages:
+
+      https://leafletjs.com/reference-1.7.1.html#map-opentooltip
     """
+    if latlng is not None:
+      self.varId = "L.marker(%s)" % latlng
     self._js.append("openTooltip()")
+    return self
+
+  def unbindTooltip(self):
+    """
+    Description:
+    -----------
+    Removes the tooltip previously bound with bindTooltip.
+
+    Related Pages:
+
+      https://leafletjs.com/reference-1.7.1.html#layer-unbindtooltip
+    """
+    self._js.append("unbindTooltip()")
+    return self
+
+  def closeTooltip(self):
+    """
+    Description:
+    -----------
+    Closes the tooltip given as parameter.
+
+    Related Pages:
+
+      https://leafletjs.com/reference-1.7.1.html#map-closetooltip
+    """
+    self._js.append("closeTooltip()")
     return self
 
   def toStr(self):
     js_fnc = ".".join(self._js)
     self._js = []
     return "%s.%s" % (self.varId, js_fnc)
+
+
+class Layer:
+  pass
 
 
 class LMarker:
@@ -35,10 +71,11 @@ class LMarker:
     """
     Description:
     -----------
+    Add the marker to the defined map.
 
     Attributes:
     ----------
-    :param selector:
+    :param selector: String. Optional. The map reference variable.
     """
     self._js.append("addTo(%s)" % selector)
     self.__is_attached = True
@@ -48,10 +85,15 @@ class LMarker:
     """
     Description:
     -----------
+    Binds a popup to all of the layers at once.
+
+    Related Pages:
+
+      https://leafletjs.com/reference-1.7.1.html#marker
 
     Attributes:
     ----------
-    :param content:
+    :param content: String. The market content.
     """
     self._js.append("bindPopup(%s)" % JsUtils.jsConvertData(content, None))
     popup = LPopup()
@@ -62,15 +104,80 @@ class LMarker:
     """
     Description:
     -----------
+    Binds a tooltip to the layer with the passed content and sets up the necessary event listeners.
+
+    Related Pages:
+
+      https://leafletjs.com/reference-1.7.1.html#layer-bindtooltip
 
     Attributes:
     ----------
-    :param content:
+    :param content: String. The tooltip content.
     """
     self._js.append("bindTooltip(%s)" % JsUtils.jsConvertData(content, None))
     tooltip = LTooltip()
     tooltip.varId = self.toStr()
     return tooltip
+
+  def setLatLng(self, latlng):
+    """
+    Description:
+    -----------
+    Changes the marker position to the given point.
+
+    https://leafletjs.com/reference-1.7.1.html#marker-setlatlng
+
+    Attributes:
+    ----------
+    :param latlng:
+    """
+    self._js.append("setLatLng(%s)" % latlng)
+    return self
+
+  def setIcon(self, icon):
+    """
+    Description:
+    -----------
+    Changes the marker icon.
+
+    https://leafletjs.com/reference-1.7.1.html#marker-seticon
+
+    Attributes:
+    ----------
+    :param icon:
+    """
+    self._js.append("setIcon(%s)" % icon)
+    return self
+
+  def setOpacity(self, opacity):
+    """
+    Description:
+    -----------
+    Changes the opacity of the marker.
+
+    https://leafletjs.com/reference-1.7.1.html#marker-setopacity
+
+    Attributes:
+    ----------
+    :param opacity:
+    """
+    self._js.append("setOpacity(%s)" % opacity)
+    return self
+
+  def setZIndexOffset(self, offset):
+    """
+    Description:
+    -----------
+    Changes the zIndex offset of the marker.
+
+    https://leafletjs.com/reference-1.7.1.html#marker-setzindexoffset
+
+    Attributes:
+    ----------
+    :param offset:
+    """
+    self._js.append("setZIndexOffset(%s)" % offset)
+    return self
 
   def toStr(self):
     if not self.__is_attached and self.leaflet_map is not None:
@@ -87,10 +194,24 @@ class LCircle(LMarker):
     self.varId = "L.%s" % shape
 
   def radius(self, num):
+    """
+    Radius of the circle, in meters.
+
+    https://leafletjs.com/reference-1.7.1.html#circle-radius
+
+    :param num:
+    """
     self.options["radius"] = num
     return self
 
   def color(self, hexcode):
+    """
+    Color of the circle.
+
+    https://leafletjs.com/reference-1.7.1.html#path
+
+    :param hexcode:
+    """
     self.options["color"] = hexcode
     return self
 
@@ -114,12 +235,16 @@ class LPopup:
     self._js, self.__is_attached = [], False
     self.varId = "L.marker(%s)" % latlng
 
-  def openPopup(self):
+  def openPopup(self, latlng=None):
     """
     Description:
     -----------
+    Opens the bound popup at the specified latlng or at the default popup anchor if no latlng is passed.
 
+    https://leafletjs.com/reference-1.7.1.html#layer-openpopup
     """
+    if latlng is not None:
+      self.setLatLng(latlng)
     self._js.append("openPopup()")
     return self
 
@@ -127,12 +252,30 @@ class LPopup:
     """
     Description:
     -----------
+    Changes the marker position to the given point.
+
+    https://leafletjs.com/reference-1.7.1.html#marker-setlatlng
 
     Attributes:
     ----------
     :param latlng:
     """
     self._js.append("setLatLng(%s)" % latlng)
+    return self
+
+  def setZIndexOffset(self, offset):
+    """
+    Description:
+    -----------
+    Changes the zIndex offset of the marker.
+
+    https://leafletjs.com/reference-1.7.1.html#marker-setzindexoffset
+
+    Attributes:
+    ----------
+    :param offset:
+    """
+    self._js.append("setZIndexOffset(%s)" % offset)
     return self
 
   def setContent(self, content):
@@ -167,14 +310,46 @@ class LtileLayer:
     self.options, self.__is_attached, self._js = options or {}, False, []
 
   def attribution(self, text):
+    """
+    The attribution control allows you to display attribution data in a small text box on a map.
+
+    :param text:
+    """
     self.options["attribution"] = text
     return self
 
+  def maxBounds(self, latLngBounds):
+    """
+    When this option is set, the map restricts the view to the given geographical bounds, bouncing the user back if the user tries to pan outside the view.
+    To set the restriction dynamically, use setMaxBounds method.
+
+    https://leafletjs.com/reference-1.7.1.html#map-maxbounds
+
+    :param latLngBounds:
+    """
+    self.options["maxBounds"] = latLngBounds
+    return self
+
   def maxZoom(self, num):
+    """
+    Maximum zoom level of the map.
+    If not specified and at least one GridLayer or TileLayer is in the map, the highest of their maxZoom options will be used instead.
+
+    https://leafletjs.com/reference-1.7.1.html#map-maxzoom
+
+    :param num:
+    """
     self.options["maxZoom"] = num
     return self
 
   def minZoom(self, num):
+    """
+    Minimum zoom level of the map. If not specified and at least one GridLayer or TileLayer is in the map, the lowest of their minZoom options will be used instead.
+
+    https://leafletjs.com/reference-1.7.1.html#map-minzoom
+
+    :param num:
+    """
     self.options["minZoom"] = num
     return self
 
@@ -183,10 +358,25 @@ class LtileLayer:
     return self
 
   def tileSize(self, num):
+    """
+    Width and height of tiles in the grid. Use a number if width and height are equal, or L.point(width, height) otherwise.
+
+    https://leafletjs.com/reference-1.7.1.html#gridlayer-tilesize
+
+    :param num:
+    """
     self.options["tileSize"] = num
     return self
 
   def zoomOffset(self, num):
+    """
+    The zoom number used in tile URLs will be offset with this value.
+
+    https://leafletjs.com/reference-1.7.1.html#tilelayer-zoomoffset
+
+    :param num:
+    :return:
+    """
     self.options["zoomOffset"] = num
     return self
 
@@ -215,6 +405,12 @@ class LMapAttributionControl:
     self.varId = varId
 
   def addAttribution(self, text):
+    """
+    The attribution control allows you to display attribution data in a small text box on a map.
+
+
+    :param text:
+    """
     return "%s.addAttribution(%s)" % (self.varId, JsUtils.jsConvertData(text, None))
 
 
@@ -251,35 +447,62 @@ class LMap:
     return self
 
   def on(self, typeEvent, jsFuncs, profile=None):
-    return JsUtils.jsWrap("%s; %s.on('%s', function(){%s})" % (
+    """
+
+    :param typeEvent:
+    :param jsFuncs:
+    :param profile:
+    """
+    return JsUtils.jsWrap("%s; %s.on('%s', function(event){%s})" % (
       self.toStr(), self.varName, typeEvent, JsUtils.jsConvertFncs(jsFuncs, toStr=True, profile=profile)))
 
   def locationfound(self, jsFuncs, profile=None):
     """
+    Fired when geolocation (using the locate method) went successfully.
 
+    https://leafletjs.com/reference-1.7.1.html#map-locationfound
     https://leafletjs.com/examples/mobile/
 
     :param jsFuncs:
     :param profile:
-    :return:
     """
     return self.on("locationfound", jsFuncs, profile)
 
   def locationerror(self, jsFuncs, profile=None):
     """
+    Fired when geolocation (using the locate method) failed.
 
     https://leafletjs.com/examples/mobile/
+    https://leafletjs.com/reference-1.7.1.html#map-locationerror
 
     :param jsFuncs:
     :param profile:
-    :return:
     """
     return self.on("locationerror", jsFuncs, profile)
 
   def createPane(self, text):
+    """
+    Stops watching location previously initiated by map.locate({watch: true}) and aborts resetting the map view
+    if map.locate was called with {setView: true}.
+
+    https://leafletjs.com/reference-1.7.1.html#map-createpane
+
+    :param text:
+    """
     return JsUtils.jsWrap("%s; %s.createPane(%s)" % (self.toStr(), self.varName, JsUtils.jsConvertData(text, None)))
 
+  def remove(self):
+    """
+    Destroys the map and clears all related event listeners.
+
+    https://leafletjs.com/reference-1.7.1.html#map-remove
+    """
+    return JsUtils.jsWrap("%s; %s.remove()" % (self.toStr(), self.varName))
+
   def fitBounds(self):
+    pass
+
+  def whenReady(self):
     pass
 
   @property
@@ -291,8 +514,26 @@ class LMap:
     return self
 
   def locate(self, options=None):
+    """
+    Tries to locate the user using the Geolocation API, firing a locationfound event with location data on success or a
+    locationerror event on failure, and optionally sets the map view to the user's location with respect to detection accuracy
+
+    https://leafletjs.com/reference-1.7.1.html#map-locate
+
+    :param options:
+    """
     options = options or {"setView": True, "maxZoom": 16}
     self._js.append("locate(%s)" % JsUtils.jsConvertData(options, None))
+    return self
+
+  def stopLocate(self):
+    """
+    Stops watching location previously initiated by map.locate({watch: true}) and aborts resetting the map view
+    if map.locate was called with {setView: true}.
+
+    https://leafletjs.com/reference-1.7.1.html#map-stoplocate
+    """
+    self._js.append("stopLocate()")
     return self
 
   def toStr(self):
@@ -348,6 +589,38 @@ class LControl:
     return "%s.%s" % (self.varId, js_fnc)
 
 
+class LEventlLatlng:
+
+  @property
+  def lat(self):
+    """
+    Return the latitude.
+
+    """
+    return "event.latlng.lat"
+
+  @property
+  def lng(self):
+    """
+    Return the longitude.
+
+    """
+    return "event.latlng.lng"
+
+  def toStr(self):
+    return "event.latlng"
+
+
+class LEvent:
+
+  @property
+  def latlng(self):
+    """
+    Get coordinate from an event on a map.
+    """
+    return LEventlLatlng()
+
+
 class LeafLet(JsPackage):
   lib_alias = {'js': "leaflet", 'css': 'leaflet'}
 
@@ -366,11 +639,14 @@ class LeafLet(JsPackage):
     return self._map
 
   def control(self, alias, options=None):
+    """
+
+    :param alias:
+    :param options:
+    """
     if alias not in self._control:
       self._control[alias] = LControl(self.src, options=options, set_var=self._selector)
     return self._control[alias]
-
-
 
   def setZoom(self, zoom):
     """
@@ -437,9 +713,20 @@ class LeafLet(JsPackage):
     return LMarker(latlng, options, leaflet_map=self.map)
 
   def circle(self, latlng, options=None):
+    """
+
+    :param latlng:
+    :param options:
+    """
     return LCircle(latlng, options, leaflet_map=self.map)
 
   def tileLayer(self, token, options=None, url="https://api.mapbox.com/styles/v1/{id}/tiles/{z}/{x}/{y}"):
+    """
+
+    :param token:
+    :param options:
+    :param url:
+    """
     return LtileLayer(self.src, token, url, options, leaflet_map=self.map)
 
   def popup(self, options=None, source=None):
@@ -493,12 +780,30 @@ class LeafLet(JsPackage):
     pass
 
   def polyline(self, latlng1, latlng2, options=None):
+    """
+
+    :param latlng1:
+    :param latlng2:
+    :param options:
+    """
     return LCircle([latlng1, latlng2], options, shape="polyline", leaflet_map=self.map)
 
   def polygon(self, latlng1, latlng2, options=None):
+    """
+
+    :param latlng1:
+    :param latlng2:
+    :param options:
+    """
     return LCircle([latlng1, latlng2], options, shape="polygon", leaflet_map=self.map)
 
   def rectangle(self, latlng1, latlng2, options=None):
+    """
+
+    :param latlng1:
+    :param latlng2:
+    :param options:
+    """
     return LCircle([latlng1, latlng2], options, shape="rectangle", leaflet_map=self.map)
 
   def circleMarker(self):